Use common configs in common functions
In (almost) every instance of using the affected common functions, the
arguments were exactly the same. Those arguments can just be moved to
be constants inside the functions, to make the calls easier to read.

To do that, the kube clients needed to be moved into the common package,
which also requires an additional "init" function. But these are
generally good things - they are common clients.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>