Correctly patch components in hosted mode tests
For the "usual" managed-cluster components, the environment variables in
the addon-controller deployment can be updated to use a newly-built
container image for the component. For the propagator, that deployment
on the hub cluster can be manually updated. For the addon-controller
itself, the make target can be "tricked" into using the new code.

Refs:
 - https://issues.redhat.com/browse/ACM-5211

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>